requirements.txt
google-adk>=0.3.0
google-cloud-aiplatform>=1.60.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
